Renamed files/classes to better reflect their purpose
