+++ dc-fs.py
@@ -4,7 +4,7 @@
 #Copyright (C) 2024 Simon Bryntse
 #License (GPL 3.0) provided in file 'LICENSE'
 
-import fs, dc
+import fs, fat
 import fuse
 import sys
 import os
@@ -59,11 +59,11 @@ def main():
         os.mkdir (temp)
 
     #Spin up system
-    discord = dc.Discord (temp, cache, channel, token, "./fat.json")
-    fuse.FUSE(fs.Filesystem(discord, cache), mount, nothreads=True, foreground=True, allow_other=False)
+    files = fat.Fat (temp, cache, channel, token, "./fat.json")
+    fuse.FUSE(fs.Filesystem(files, cache), mount, nothreads=True, foreground=True, allow_other=False)
 
     #Gracefully shut down after unmount
-    discord.exit ()
+    files.exit ()
 
     try:
         os.rmdir (cache)
